+++ src/routes/graphql/schemas/mutations/index.ts
@@ -1,0 +1,3 @@
+export * from "./create-user.mutation";
+export * from "./create-post.mutation";
+export * from "./create-profile.mutation";
